requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.5.1
pydantic-settings==2.0.3
python-multipart==0.0.6
aiofiles==23.1.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.4.0
requests==2.31.0
elasticsearch==8.12.0
sentence-transformers==2.5.1
hdbscan==0.8.32
celery==5.3.0
redis==4.6.0
openai==1.12.0
numpy==1.26.4
scipy==1.11.4
google-generativeai==0.3.2
python-dotenv==1.0.0